Checking for packer existence.

+++ scripts/travis/bundle_install.sh
@@ -7,11 +7,9 @@ cd ~
 wget -O "remote_$FILE_NAME" "https://$AWS_S3_BUCKET.s3.amazonaws.com/$FILE_NAME" && tar -xf "remote_$FILE_NAME"
 wget -O "remote_$FILE_NAME.sha2" "https://$AWS_S3_BUCKET.s3.amazonaws.com/$FILE_NAME.sha2"
 
-
-
-
-####### only if file doesn't exist
-wget --no-check-certificate https://dl.bintray.com/mitchellh/packer/$PACKER_FILENAME && unzip -d packer $PACKER_FILENAME
-mv packer .bundle
+if [ ! -f $PWD/.bundle/packer/packer ] then
+  wget --no-check-certificate https://dl.bintray.com/mitchellh/packer/$PACKER_FILENAME && unzip -d packer $PACKER_FILENAME
+  mv packer .bundle
+fi
 
 exit 0
